refactor(tui): split ui into a separate mod from event list
